Merge into my master project (#10)
* Add Random Weather Generator

Add a hidden tab on Main Window for a random weather generator.  This
generates weather roughly based on the srd gamemastering environment
weather web page.

* Add some more colors to Color menu

Add some more colors besides Red and None to the color table.  Added
Green, Yellow and Blue. Changed from uint hex value to use
System.Drawing.Color

* Power Attack feat condition.  Able to add a Power Attack condition to monsters.  Updates melee weapons. Does not account for single handed weapon used with two hands

* Updated Monster DB to allow the addition of multiple monsters.  The number
is clamped between 1 and 10 for now.  The value will reset back to 1 once
the monster is added. Need to fix the font size in the textbox to it is
smaller and fits better.

* Added single validation function.  Updated layout to auto size textbox and
set font size to fit.